refactor: combine btns in one component
