+++ packages/client/src/ui/button/index.tsx
@@ -1,13 +1,38 @@
 import cn from "classnames";
 import style from "./style.module.scss";
 
-type buttonType = {
+type TProps = {
   name: string;
   type?: "submit" | "reset" | "button";
   disabled?: boolean;
+  transparent?: boolean;
+  dangerous?: boolean;
   red?: boolean;
   yellow?: boolean;
   onClick?: (event: React.MouseEvent) => void;
+};
+
+const FlatButton = ({
+  name,
+  type = "button",
+  disabled,
+  transparent,
+  dangerous,
+  onClick,
+}: TProps) => {
+  return (
+    <button
+      className={cn(style.button, {
+        [style.transparent]: transparent,
+        [style.dangerous]: dangerous,
+      })}
+      type={type}
+      disabled={disabled}
+      onClick={onClick}
+    >
+      {name}
+    </button>
+  );
 };
 
 const RoundButton = ({
@@ -17,7 +42,7 @@ const RoundButton = ({
   red,
   yellow,
   onClick,
-}: buttonType) => {
+}: TProps) => {
   return (
     <div
       className={style["button-el"]}
@@ -40,4 +65,9 @@ const RoundButton = ({
   );
 };
 
-export default RoundButton;
+const Button = {
+  Flat: FlatButton,
+  Round: RoundButton,
+};
+
+export default Button;
