--- packages/client/src/ui/flatButton/style.module.scss
@@ -1,48 +1,0 @@
-.button {
-  position: relative;
-  width: 280px;
-  height: 37px;
-  border-radius: 8px;
-  color: var(--white);
-  background-color: var(--primary);
-  transition:
-    box-shadow var(--transition-bezier),
-    background-color var(--transition-bezier);
-
-  &.transparent {
-    background-color: transparent;
-    color: var(--primary);
-    transition: text-shadow var(--transition-bezier);
-  }
-
-  &.dangerous {
-    background-color: var(--danger);
-  }
-
-  &:hover {
-    box-shadow: 0px 0px 5px 0px var(--shadow-strong);
-
-    &.transparent {
-      box-shadow: 0px 0px 0px 0px var(--black);
-      text-shadow: 0px 0px 0.7px var(--shadow-strong);
-    }
-  }
-
-  &:active {
-    box-shadow: 0px 0px 1px 0px var(--shadow-strong);
-
-    &.transparent {
-      box-shadow: 0px 0px 0px 0px var(--black);
-      text-shadow: 0px 0px 0.2px var(--shadow-strong);
-    }
-  }
-
-  &:disabled {
-    background-color: var(--border);
-    cursor: not-allowed;
-
-    &:hover {
-      box-shadow: 0px 0px 0px 0px var(--black);
-    }
-  }
-}
